% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decompose}
\alias{decompose}
\title{Decompose Model}
\usage{
decompose(m, xes, ges, i = 1, j = 0)
}
\arguments{
\item{m}{a linear regression model}

\item{xes}{a character vector giving the names of the independent (or control) variables}

\item{ges}{a character vector giving the name(s) of the grouping variable and any polynomial transformations. It is assumed that the first element is the original grouping variable.}

\item{i, j}{values of the group variable to compare}
}
\value{
Returns a list with the following items:
  \code{model} - a copy of the \code{m} object passed to the function
  \code{total} - the total difference for the outcome variable
  \code{explained} - the amount of difference attributed to the explained part
  \code{unexplained} - the amount of difference attributed to the unexplained part
}
\description{
This function takes a model, and list of variables, to generate a decomposition
of the difference in outcome for the grouping variable.
}
