% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{estimateX}
\alias{estimateX}
\title{This function calculates the average value of x at values of g=i}
\usage{
estimateX(x, g, i = 0)
}
\description{
This function calculates the average value of x at values of g=i
}
