% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{gen_decomposed_results_multinom}
\alias{gen_decomposed_results_multinom}
\title{Decompose results of a mulitnomial regression model}
\usage{
gen_decomposed_results_multinom(
  fitted.mdl,
  df.analysis,
  list_of_IVs,
  list_of_grouping_vars,
  values_of_grouping_vars
)
}
\arguments{
\item{fitted.mdl}{A multinom regression object from the nnet package}

\item{df.analysis}{the dataframe used to run the regression model}

\item{list_of_IVs}{A character vector indicator the predictor variables included in the probit model}

\item{list_of_grouping_vars}{A character vector of grouping variables included in the probit model}

\item{values_of_grouping_vars}{A vector containing all possible values of the grouping variable.}
}
\value{
A list containing two objects. One, the fitted OLS model based
  on the estimated log-odds continuous variable. Two, a data frame containing the decomposition
  comparing every possible combination of values for the grouping variable.
}
\description{
Given a properly expressed multinom regression, decompose the results into an
explained and unexplained parts.
}
