% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{gen_decomposed_results_probit}
\alias{gen_decomposed_results_probit}
\title{Given a properly expressed probit regression, decompose the results into an
explained and unexplained parts.}
\usage{
gen_decomposed_results_probit(
  fitted.mdl,
  list_of_IVs,
  list_of_grouping_vars,
  values_of_grouping_vars
)
}
\arguments{
\item{fitted.mdl}{A probit regression object}

\item{list_of_IVs}{A character vector indicator the predictor variables included in the probit model}

\item{list_of_grouping_vars}{A character vector of grouping variables included in the probit model}

\item{values_of_grouping_vars}{A vector containing all possible values of the grouping variable.}
}
\value{
A list containing two objects. One, the fitted OLS model based
  on the estimated log-odds continuous variable. Two, a data frame containing the decomposition
  comparing every possible combination of values for the grouping variable.
  @export
}
\description{
Given a properly expressed probit regression, decompose the results into an
explained and unexplained parts.
}
