% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{getUnexplained}
\alias{getUnexplained}
\title{Get unexplained part of KOB decomposition}
\usage{
getUnexplained(m, xes, ges, i = 1, j = 0)
}
\arguments{
\item{m}{a correctly specified model for an KOB decomposition}

\item{xes}{a character vector giving the names of the independent (or control) variables}

\item{ges}{a character vector giving the name(s) of the grouping variable and any polynomial transformations. It is assumed that the first element is the original grouping variable.}

\item{i, j}{values of the group variable to compare}
}
\value{
The amount of difference attributable to the unexplained part of a KOB decomposition.
}
\description{
This function calculates the unexplained part of a Kitawaga-Oaxaca-Blinder (KOB) decomposition
generalized to a continuous grouping variable. This function is rarely directly used by the user.
}
