% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{simMlogitRes}
\alias{simMlogitRes}
\title{Generate simulated residuals for a latent variable for mlogit}
\usage{
simMlogitRes(m, df, outcome = "y")
}
\arguments{
\item{m}{a multinom regression model from the nnet package}

\item{df}{the data used to run the probit regression}

\item{outcome}{the name of the outcome variable}
}
\value{
A matrix of simulated residuals, one column for each possible outcome
}
\description{
Based on Wolff (2012), this will generate simulated residuals from a mlogit model
such that the generated residual plus the linear prediction corresponds to the observed
outcome
}
