% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{simulateResiduals}
\alias{simulateResiduals}
\title{Generate simulated residuals for a latent variable}
\usage{
simulateResiduals(m, df, outcome = "y")
}
\arguments{
\item{m}{a probit regression model}

\item{df}{the data used to run the probit regression}

\item{outcome}{the name of the outcome variable}
}
\value{
A vector of simulated residuals
}
\description{
Based on Wolff (2012), this will generate simulated residuals from a probit model
such that the generated residual plus the linear prediction corresponds to the observed
outcome (i.e. Y==1:Y*>0; Y==0:Y*<=0)
}
